# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This documentation for One-Eyed-Willy pdf generator.",
        "title": "One-Eyed-Willy pdf generation API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Aníbal Deboni Neto",
            "email": "anibaldeboni@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/pdf/generate": {
            "post": {
                "description": "Generate a new pdf file from a html string",
                "consumes": [
                    "applcation/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "pdf"
                ],
                "summary": "Create a pdf",
                "parameters": [
                    {
                        "description": "Base64 encoded string of a html",
                        "name": "html",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createPdfFromHTMLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
        "/pdf/merge": {
            "post": {
                "description": "Merges two or more pdfs",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "pdf"
                ],
                "summary": "Merge pdfs",
                "parameters": [
                    {
                        "type": "file",
                        "description": "pdf files to merge",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.createPdfFromHTMLRequest": {
            "type": "object",
            "required": [
                "html"
            ],
            "properties": {
                "html": {
                    "type": "string"
                }
            }
        },
        "utils.Error": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        }
    }
}